nisria-backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
cron-descriptor==1.4.5
Django==5.2.1
django-celery-beat==2.8.1
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.5.3
Markdown==3.8
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
